reorg to use multiple different datasets. added quilt functionality
