+++ .github/workflows/test-and-lint.yml
@@ -25,6 +25,8 @@ jobs:
       run: |
         pytest --cov-report xml --cov=fov_processing_pipeline fov_processing_pipeline/tests/
         codecov -t ${{ secrets.CODECOV_TOKEN }}
+      env: |
+        CI=True
 
   lint:
     runs-on: ubuntu-latest
