+++ fov_processing_pipeline/tests/data/test_utils.py
@@ -1,10 +1,10 @@
-from .. import data
 import numpy as np
 import pandas as pd
 
+from ...data import utils
 
 def test_cell_data_to_fov_data(demo_cell_data):
-    fov_data = data._cell_data_to_fov_data(demo_cell_data)
+    fov_data = utils.cell_data_to_fov_data(demo_cell_data)
 
     # make sure the length of the FOV dataframe is the same length as the number of unique FOVIds
     assert len(np.unique(fov_data["FOVId"])) == len(fov_data["FOVId"])
@@ -27,19 +27,19 @@ def test_trim_data(demo_fov_data):
     demo_multi_fov_data["FOVId_rng"] = range(len(protein_list)*2)
 
     # check that trimming to one cell line works
-    trim_df = data.trim_data(demo_multi_fov_data, ["Fibrillarin"])
+    trim_df = utils.trim_data(demo_multi_fov_data, ["Fibrillarin"])
     assert len(pd.unique(trim_df["ProteinDisplayName"])) == 1
 
     # check that trimming to a smaller number of fovs works
-    trim_df = data.trim_data(demo_multi_fov_data, ["Fibrillarin"], 1)
+    trim_df = utils.trim_data(demo_multi_fov_data, ["Fibrillarin"], 1)
     assert trim_df.shape[0] == 1
 
     # check that trimming to a cell line list that is the same as the existing list does nothing
-    trim_df = data.trim_data(demo_multi_fov_data, protein_list)
+    trim_df = utils.trim_data(demo_multi_fov_data, protein_list)
     assert trim_df.shape[0] == demo_multi_fov_data.shape[0]
 
     # check that trimming to a cell line not in the list leaves nothing
-    trim_df = data.trim_data(demo_multi_fov_data, ["Nonexistent protein name"])
+    trim_df = utils.trim_data(demo_multi_fov_data, ["Nonexistent protein name"])
     assert trim_df.shape[0] == 0
 
 
@@ -61,9 +61,9 @@ def test_trim_data_by_cellline_fov(demo_fov_data):
     demo_multi_fov_data["FOVId_rng"] = range(len(protein_list)*2)
 
     # check that trimming to a smaller number of fovs works
-    trim_df = data.trim_data_by_cellline_fov_count(demo_multi_fov_data, 1)
+    trim_df = utils.trim_data_by_cellline_fov_count(demo_multi_fov_data, 1)
     assert trim_df.shape[0] == len(protein_list)
 
     # check that trimming to a larger number of fovs leaves it untouched
-    trim_df = data.trim_data_by_cellline_fov_count(demo_multi_fov_data, 5)
+    trim_df = utils.trim_data_by_cellline_fov_count(demo_multi_fov_data, 5)
     assert trim_df.shape[0] == demo_multi_fov_data.shape[0]
